big restructure to submodule structure per registry
