+++ openeo_fastapi/client/collections/__init__.py
@@ -1,0 +1,2 @@
+from openeo_fastapi.client.collections.models import *
+from openeo_fastapi.client.collections.register import CollectionRegister
