+++ openeo_fastapi/client/jobs/__init__.py
@@ -1,0 +1,2 @@
+from openeo_fastapi.client.jobs.models import *
+from openeo_fastapi.client.jobs.register import JobsRegister
