+++ tests/data/fake-job-post.json
@@ -7,7 +7,7 @@
         "deprecated":false,
         "exceptions":{},
         "experimental":false,
-        "id":"0eoBMjx6ng1BvD7C",
+        "id":"",
         "process_graph":{
             "loadcollection1":{
                 "process_id": "load_collection",
